rnn/dcrnn_train.py
from __future__ import absolute_import
from __future__ import division
from __future__ import print_function

import argparse
import yaml

from rnn.lib.utils import load_graph_data
from rnn.model.dcrnn_supervisor import DCRNNSupervisor


def main(args):
    with open(args.config_filename) as f:
        supervisor_config = yaml.load(f)
        supervisor_config['model']['adaptive_supports'] = args.adaptive_supports

        graph_pkl_filename = supervisor_config['data'].get('graph_pkl_filename')
        sensor_ids, sensor_id_to_ind, adj_mx = load_graph_data(graph_pkl_filename)

        supervisor = DCRNNSupervisor(adj_mx=adj_mx, **supervisor_config)

        supervisor.train()


if __name__ == '__main__':
    parser = argparse.ArgumentParser()
    parser.add_argument('--config_filename', default=None, type=str,
                        help='Configuration filename for restoring the model.')
    parser.add_argument('--use_cpu_only', default=False, type=bool, help='Set to true to only use cpu.')
    parser.add_argument('--adaptive_supports', default=0)
    args = parser.parse_args()
    main(args)
